Initial movement of todos into its own cog
